requirements.txt
catboost==1.0.4
Django==4.0.5
django_rest_knox==4.2.0
djangorestframework==3.13.1
knox==0.1.14
numpy==1.22.2
scikit_learn==1.1.1
gunicorn==20.0.4
psycopg2==2.8.5
django-cors-headers==3.13.0
scipy==1.8.1
pandas==1.3.5

